Propagate module override recursively in generic signatures
Signatures of generic types that contain multiple levels of generics
currently apply module override only on the level where it was defined.
This complicates cases when we copy a signature as-is from one module
and want to apply module override to it since we store it in another
module. We would need to parse the whole signature and insert overrides
at all levels.
This change modifies runtime so that the module override is propagated
recursively through a generic signature hierarchy. A related change
was necessary on the crossgen side.