Changes to make inlined pinvoke codegen look more like the fragile NI codegen, for better performance.
These changes are *only* for System.Private.CoreLib, and any other assembly within the same bubble. The assumption
here is that assemblies in this bubble will always ship alongside a CoreCLR runtime, so it's ok to have fragile NI
codegen for these scenarios since we'll always have compatible runtime bits.