Delete suspicious code in ZapInfo::getCallInfo
We abort R2R compiling methods with `thisTransform == CORINFO_BOX_THIS`. This means we don't R2R compile some methods that do virtual calls on valuetypes (e.g. calling `ToString` on a struct that doesn't itself provide a `ToString` method).

I don't understand why though. CORINFO_BOX_THIS is an instruction to the codegen to box the valuetype before doing the virtual call. We shouldn't need a "boxing stub" as the comment I'm deleting implies because the valuetype is already boxed by codegen. This should be handled as a normal virtual call.