Allow generating HW intrinsics in crossgen
We currently don't precompile methods that use hardware intrinsics because we don't know the CPU that the generated code will run on. Jitting these methods slows down startup and accounts for 3% of startup time in PowerShell.

With this change, we're going to lift this restriction for CoreLib (the thing that matters for startup) and support generating HW intrinsics for our minimum supported target ISA (SSE/SSE2). This means that code that uses intrinsics higher than this will be compiled with IsSupported = false.

We will rely on tiered JITting to eventually replace the pregenerated method body that targets the minimum supported ISA with the most efficient code supported by the processor.