Use getrandom() if available to implement PAL_Random()
This system call has been added to Linux in 3.17.  The main benefit
of it is not requiring /dev/urandom to exist -- which might be the
case if running under a chroot where /dev isn't present.  It also
does not require another file descriptor.  It's otherwise equivalent
to reading from /dev/urandom (if passing the GRND_NONBLOCK flag).