Introduce a "PerfScore" rating in the JIT disassembly for each method This rating that estimated the execution costs of each instruction and also takes into accouunt the BasicBlock weights and yeilds a single score that can be used to judge the overall execution throughput and code quality of the generated code.
THe inlining of methods will cause a score to increase, so this can't be used to directly compare
optimized vs. non-optimixed code quality.  We are not modeling instruction scheduling or the
exact latency of memory reads feeding into their next use.  Instead we just model a simplistic
fixed read latency, with a small benefit for reads from the stack.

Estimates for the x64 instruction throughput are based upon data for the Intel Skylake architecture.