Fix typo: `_TARGET_X86` -> `_TARGET_X86_`
This typo was in #24989 so would be a new regression in 3.0.
In an x86 build, it causes us to not get the cache size correct,
leading us to use a smaller default cache size and do more GCs.

Tested with GCPerfSim and this PR reduces TotalNumberGCs by 33% using an x86 build.