Dictionary optimization after deserialization
The optimization in `Dictionary<,>` around `_comparer` will be
nullified if it's created via deserialization.

If `_comparer` is `null`, the comparer will be serialized using
`EqualityComparer<T>.Default`, so when it's being deserialized,
it should be assigned back to null to run into the faster branch
when lookup.

Fix dotnet/corefx#40150