Fix faulty assert in card_transition.
I fixed an off-by-one issue in find_card, and when I switched workstation GC back to no card marking stealing, this assert at the end of card_transition fired in the case where limit, end and card_address(end_card) all coincide, because we access the card for the end address which we really shouldn't look at.

The fix is simply to compare limit to end instead of card_address(end_card).