Address Egor's feedback
1) Renamed build-test.yml to build-test-job.yml;

2) Renamed run-test.yml to run-test-job.yml;

3) Removed passing crossrootfsDir to run-test-job from test-job;

4) Added new job template xplat-test-job.yml where I moved the
test-specific properties I originally put into xplat-job.yml and
Egor didn't like them there;

5) Added new step template upload-artifact-step.yml to zip up
and publish a given subtree under a given artifact name. Tried
to design its API so that it should be usable for zipping the entire
GitHub enlistment should we choose to do that. Changed
build-test-job.yml to call the step.

6) Added new step template download-artifact-step.yml to download
and unzip a given subtree from a given artifact. Tried to design
its API so that it should be usable for restoring the GitHub
enlistment in the various jobs within a particular pipeline.

Thanks

Tomas