Extend Pri1 test build timeout
According to my latest coreci-outerloop runs, Pri1 test build
jobs typically take an hour and a quarter or so and the 90-minute
timeout is in most cases insufficient to finish the job including
zipping and artifact upload to Azure. I have tentatively increased
test build timeout for Pri1 tests to 2 hours.

Thanks

Tomas