Fix issue with locals overlapping out of scope GCFrame
More aggressive C/C++ optimizations done by VS2019 are breaking fragile
assumptions of the CoreCLR "manually managed code".

Unwinding of Frame chains accesses stack local variables after the stack
frame has been unwound, but it depends on their content to be left
intact. The new compiler is breaking this assumption by stack-packing a
different variable over it.

This change fixes the problem by adding a destructor to GCFrame that
pops the frame from the per-thread Frame list.

I also had to refactor two functions where the compiler was complaining
about mixing SEH and C++ EH in single function.

With these changes applied, there was still a problem that I've
discovered when running CoreCLR tests with GCStress 3. When the
ExceptionTracker::m_pInitialExplicitFrame was still pointing to a frame
that was already removed from the explicit Frame chain. When the
ExceptionTracker::HasFrameBeenUnwoundByAnyActiveException was walking
the frame chain starting at m_pInitialExplicitFrame to figure out if a given
frame was already unwound, it has walked to the destryoyed GCFrame and
crashed.
After various attempts to update the m_pInitialExplicitFrame during the
exception handling so that the explicit frames chain starting at it
doesn't contain the destroyed GC frames, I have decided to drop the
m_pInitialExplicitFrame
completely and rather add a flag to the Frame indicating whether it was
already unwound by an exception or not. The Frame already had a method
ExceptionUnwind
that is called when the unwinding happens, so I could just set the flag
in there.
Due to this change, I had to fix a couple of ASM helpers and some places
in native code generators in the runtime. There were places where we were
using numeric constants for various offsets and sizes instead of constants
defined in asmconstants.h and verified at build time.