Update null ref annotation for object.Equals(object?)
The `object.Equals(object)` method accepts a null argument, but if the method returns `true`, it's a guarantee to the caller that the argument is *not* null, since an instance can never be equal to a null reference. This annotation reflects this guarantee so the compiler can not generate a warning when `#nullable` features are used.

Granted, someone could override `Equals` in their own type and make it return `true` even if the argument is `null`. But that would be breaking the contract. And this annotation won't break them either, but it's a great help for the expected case.