WIP: refactor osGroup / osIdentifier pairs to osGroup / osSubgroup
Based on my experiments with the change to build managed *nix test
artifacts on OSX I believe we're hitting the template expansion
limit very narrowly. For this reason I'm experimenting with reducing
the expansion tree by switching over osIdentifier to be a straight
concatenation of osGroup and osSubgroup. I believe it has the
potential to reduce the expansion size and it cleans up the scripts
in various places.

Thanks

Tomas