Add ability to define inline labels
Change IGF_EMIT_ADD to IGF_EXTEND, which conceptually indicates that
this instruction groups extends the previous. As before, this means that
the GC info is inherited from the previous instruction group. Use this
mechanism in a new genDefineInlineTempLabel, so that labels generated
after calls duet GT_LABEL nodes have the proper GC info.