Fix build on RHEL 6
There is a bug in the ld linker on RHEL 6 that causes linking error when
files are compiled with -ffunction-sections and -fdata-sections options
and executables linked with -pie and --gc-sections options.
This change is a workaround that doesn't add the -pie linker option
if building on RHEL / CentOS 6 with corebundle build enabled.
The downside is that the corebundle binary is not ASLR enabled in
that case.